Remove ImageFeedCard
