+++ client/src/pages/UserPage.tsx
@@ -21,7 +21,6 @@ import {decodeToken} from "react-jwt";
 import CircularProgress from '@mui/material/CircularProgress';
 import axios from "axios";
 import {base_be_url} from "../util/constants";
-import ImageFeedCard from "../components/ImageFeedCard";
 import Typography from "@mui/material/Typography";
 import SuggestedUserCard from "../components/SuggestedUserCard";
 
@@ -194,7 +193,7 @@ const UserPage = () => {
                 <div className="mt-5 grid lg:grid-cols-4 md:grid-cols-3 sm:grid-cols-2 gap-5 p-10 grid-cols-1 mx-[10vw]">
                 {userData.images.map((image: any) => (
                         <div key={image.id} className="mt-2">
-                            <ImageCard imageData={image} />
+                            <ImageCard isFeed={false} imageData={image} />
                         </div>
                         )
                     )}
@@ -209,7 +208,7 @@ const UserPage = () => {
                 <div className="mt-5 grid lg:grid-cols-4 md:grid-cols-3 sm:grid-cols-2 gap-5 p-10 grid-cols-1 mx-[10vw]">
                     {loggedInUsername === username && option === 2 && userData.feedImages.map((imageObj: any)=> (
                         <div className="mt-2">
-                                <ImageFeedCard imageData={imageObj} />
+                            <ImageCard isFeed={true} imageData={imageObj} />
                         </div>
                     ))}
                 </div>
